requirements.txt
absl-py==1.3.0
addict==2.4.0
anyio==3.6.1
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644569681205/work
async-generator==1.10
attrs==22.1.0
Babel==2.10.3
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
black==22.6.0
bleach==5.0.1
Brotli==1.0.9
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
chamfer==2.0.0
chardet==5.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
commonmark==0.9.1
ConfigArgParse==1.5.3
configobj==5.0.6
contextvars==2.4
cryptography @ file:///croot/cryptography_1665612644927/work
cubic-feature-sampling==1.1.0
cycler==0.11.0
Cython==0.29.32
dash==2.6.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.3
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
deprecation==2.1.0
descartes==1.1.0
docker-pycreds==0.4.0
DracoPy==1.2.0
emd-ext==0.0.0
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
exceptiongroup==1.1.0
fastjsonschema @ file:///opt/conda/conda-bld/python-fastjsonschema_1661371079312/work
filelock==3.4.1
fire==0.4.0
flake8==5.0.4
flake8-bugbear==22.7.1
flake8-comprehensions==2.3.0
Flask==2.2.2
Flask-Compress==1.13
future==0.18.2
fvcore @ file:///home/conda/feedstock_root/build_artifacts/fvcore_1652361911092/work
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.15.0
google-auth-oauthlib==0.4.6
gridding==2.1.0
gridding-distance==1.0.0
grpcio==1.51.1
h5py==3.7.0
hausdorff==0.2.6
hydra-core==0.11.3
idna @ file:///croot/idna_1666125576474/work
imageio==2.22.3
immutables==0.18
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562407465/work
importlib-resources==5.4.0
iniconfig==2.0.0
iopath==0.1.9
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.6.6
isort==4.3.21
itsdangerous==2.1.2
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.1.0
json5==0.9.9
jsonschema @ file:///opt/conda/conda-bld/jsonschema_1663375472438/work
jupyter @ file:///tmp/abs_33h4eoipez/croots/recipe/jupyter_1659349046347/work
jupyter-console==6.4.4
jupyter-packaging==0.10.6
jupyter-server @ file:///tmp/abs_b88b31b8-83b9-476d-a46d-e563c421f38fvsnyi1ur/croots/recipe/jupyter_server_1658754481507/work
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab @ file:///tmp/abs_12f3h01vmy/croots/recipe/jupyterlab_1658907535764/work
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jupyterlab_server @ file:///opt/conda/conda-bld/jupyterlab_server_1664908231281/work
-e git+https://github.com/NVIDIAGameWorks/kaolin.git@c537bf76111153b83555787354b6f8f20d4d58a0#egg=kaolin
kiwisolver==1.3.1
llvmlite==0.36.0
lmdb==1.4.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.3.4
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mccabe==0.7.0
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
msgpack==1.0.4
msgpack-numpy==0.4.8
mypy-extensions==0.4.3
nbclassic @ file:///opt/conda/conda-bld/nbclassic_1644943264176/work
nbclient==0.7.0
nbconvert==6.5.4
nbformat @ file:///opt/conda/conda-bld/nbformat_1663744952973/work
nest-asyncio==1.5.6
networkx==2.6.3
nose==1.3.7
notebook @ file:///tmp/abs_abf6xa6h6f/croots/recipe/notebook_1659083654985/work
numba==0.53.1
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
nuscenes-devkit==1.1.9
oauthlib==3.2.2
olefile==0.46
omegaconf==1.4.1
open3d==0.16.0
opencv-python==4.6.0.66
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.1.5
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathspec==0.9.0
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.3.0
pkgutil_resolve_name @ file:///opt/conda/conda-bld/pkgutil-resolve-name_1661463321198/work
platformdirs==2.4.0
plotly==5.11.0
pluggy==1.0.0
ply==3.11
plyfile==0.7.4
-e git+https://github.com/erikwijmans/Pointnet2_PyTorch.git@b5ceb6d9ca0467ea34beb81023f96ee82228f626#egg=pointnet2
pointnet2-ops @ file:///home/skewen/lidar-base-point-cloud-error-concealment/Pointnet2_PyTorch/pointnet2_ops_lib
portalocker==1.4.0
pptk==0.1.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil==5.9.1
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycodestyle==2.9.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyflakes==2.5.0
Pygments==2.13.0
pymesh==1.0.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyquaternion==0.9.9
pyrsistent==0.19.1
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
pytest==7.2.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-lightning==0.7.1
PyTorch-metric==0.0.0
pytorch3d==0.3.0
pytz @ file:///opt/conda/conda-bld/pytz_1654762638606/work
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.2.1
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-image==0.16.2
scikit-learn==0.24.2
scipy==1.4.1
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sentry-sdk==1.9.4
setproctitle==1.2.3
Shapely==1.8.2
shortuuid==1.0.9
sim2net==3.1.2
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smmap==5.0.0
sniffio @ file:///tmp/build/80754af9/sniffio_1614030478629/work
soupsieve @ file:///croot/soupsieve_1666296392845/work
tabulate @ file:///home/conda/feedstock_root/build_artifacts/tabulate_1665138452165/work
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1667338503774/work
terminado @ file:///tmp/build/80754af9/terminado_1644304459949/work
testpath @ file:///opt/conda/conda-bld/testpath_1655908557405/work
threadpoolctl==3.1.0
tinycss2==1.2.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==1.2.3
tomlkit==0.11.4
torch==1.4.0
torchvision==0.5.0
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1662214488106/work
traitlets==5.5.0
trimesh==3.13.1
typed-ast==1.5.4
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///croot/urllib3_1666298941550/work
wandb==0.13.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.2.2
widgetsnbextension @ file:///tmp/build/80754af9/widgetsnbextension_1645009354203/work
yacs @ file:///home/conda/feedstock_root/build_artifacts/yacs_1645705974477/work
zipp @ file:///opt/conda/conda-bld/zipp_1652341764480/work
